:truck: : security 폴더 common 폴더로 move
