+++ Match-Api/src/main/java/com/example/matchapi/common/security/JwtAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package com.example.matchapi.security;
+package com.example.matchapi.common.security;
 
 import com.example.matchdomain.user.exception.UserAuthErrorCode;
 import org.json.JSONException;
