+++ Match-Api/src/main/java/com/example/matchapi/config/SecurityConfig.java
@@ -1,12 +1,11 @@
 package com.example.matchapi.config;
 
+import com.example.matchapi.common.security.*;
 import com.example.matchapi.security.*;
 import com.example.matchdomain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.builders.WebSecurity;
@@ -17,7 +16,6 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.cors.CorsUtils;
-import org.springframework.http.HttpMethod;
 
 @EnableWebSecurity
 @EnableGlobalMethodSecurity(prePostEnabled = true)
